gcf/end-point-requirement/src/requirements.txt
functions-framework == 3.3.0
pandas == 2.0.0
SQLAlchemy == 2.0.9
psycopg2-binary == 2.9.6
google-cloud-secret-manager == 2.16.1
Flask == 2.2.3
